added repository pattern
added ExceptionMiddleware with redirect to CrashActivity.java

altered primary keys to become long

Implemented Paging using PagingHelper

Implemented an Executor Singleton using ExecutorHelper

+++ Android/app/src/main/java/com/fivegearszerochill/noted/room/entity/NoteCategoryEntity.java
@@ -9,7 +9,7 @@ public class NoteCategoryEntity {
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "category_id")
-    private int categoryId;
+    private long categoryId;
 
     @ColumnInfo(name = "description")
     private String description;
@@ -26,7 +26,7 @@ public class NoteCategoryEntity {
         this.categoryId = categoryId;
     }
 
-    public int getCategoryId() {
+    public long getCategoryId() {
         return categoryId;
     }
